refactor: harden services and clean legacy handoff
Made-with: Cursor
